bake: new hclparser package
